+++ src/aqua_diagnostics/core/__init__.py
@@ -2,13 +2,10 @@ from .diagnostic import Diagnostic
 from .time_util import start_end_dates, round_startdate, round_enddate
 from .util import template_parse_arguments, open_cluster, close_cluster
 from .util import load_diagnostic_config, merge_config_args
-from .util import convert_data_units
-from .histogram import histogram
 from .output_saver import OutputSaver
 
 __all__ = ['Diagnostic',
            'start_end_dates', 'round_startdate', 'round_enddate',
            'template_parse_arguments', 'open_cluster', 'close_cluster',
            'load_diagnostic_config', 'merge_config_args',
-           'convert_data_units', 'histogram',
            'OutputSaver']
