+++ src/aqua_diagnostics/core/util.py
@@ -155,42 +155,3 @@ def merge_config_args(config: dict, args: argparse.Namespace,
             logger.debug(f"  - {ref['catalog']} {ref['model']} {ref['exp']} {ref['source']}")
 
     return config
-
-
-def convert_data_units(data, var: str, units: str, loglevel: str = 'WARNING'):
-    """
-    Make sure that the data is in the correct units.
-
-    Args:
-        data (xarray Dataset or DataArray): The data to be checked.
-        var (str): The variable to be checked.
-        units (str): The units to be checked.
-    """
-    logger = log_configure(log_name='check_data', log_level=loglevel)
-
-    data_to_fix = data[var] if isinstance(data, xr.Dataset) else data
-    final_units = units
-    initial_units = data_to_fix.units
-
-    conversion = convert_units(initial_units, final_units)
-
-    factor = conversion.get('factor', 1)
-    offset = conversion.get('offset', 0)
-
-    if factor != 1 or offset != 0:
-        logger.debug('Converting %s from %s to %s',
-                     var, initial_units, final_units)
-        data_to_fix = data_to_fix * factor + offset
-        data_to_fix.attrs['units'] = final_units
-        log_history(data_to_fix, f"Converting units of {var}: from {initial_units} to {final_units}")
-    else:
-        logger.debug('Units of %s are already in %s', var, final_units)
-        return data
-
-    if isinstance(data, xr.Dataset):
-        data_fixed = data.copy()
-        data_fixed[var] = data_to_fix
-    else:
-        data_fixed = data_to_fix
-
-    return data_fixed
