+++ src/aqua_diagnostics/global_biases/global_biases.py
@@ -1,8 +1,8 @@
 import pandas as pd
 import xarray as xr
 from aqua.logger import log_configure
-from aqua.diagnostics.core import Diagnostic, convert_data_units
-from aqua.util import select_season
+from aqua.util import select_season, convert_data_units
+from aqua.diagnostics.core import Diagnostic
 from aqua.exceptions import NoDataError
 from .util import handle_pressure_level
